Address timestamp casting issue
Address bizarre error:
TypeError: Cannot cast array data from dtype('float64') to dtype('int64') according to the rule 'safe'
2017-03-06 11:28:30.825Z - time_plot.py:948 - WARNING - Exception in RingBufferProcess: Cannot cast array data from dtype('float64') to dtype('int64') according to the rule 'safe'
Traceback (most recent call last):
  File "/home/kat/ve/bin/time_plot.py", line 351, in RingBufferProcess
    ts = datasd.select_timeseriesdata(product=0, start_time=0, end_time=1e100, include_ts=True)[0]#gets all timestamps only
  File "/home/kat/ve/local/lib/python2.7/site-packages/katsdpdisp/data.py", line 2198, in select_timeseriesdata
    rolled_ts = np.roll(self.storage.timeseriests,-roll_point)
  File "/home/kat/ve/local/lib/python2.7/site-packages/numpy/core/numeric.py", line 1386, in roll
    res = a.take(indexes, axis)

+++ katsdpdisp/data.py
@@ -1092,7 +1092,7 @@ class SpeadSDReceiver(threading.Thread):
                         bls_ordering_version = self.ig['bls_ordering'].version
                     hasdata = (self.ig['sd_data'].value is not None)
                     if (hasdata) or (self.ig['sd_percspectrum'].value is not None):
-                        ts = self.ig['sd_timestamp'].value * 10.0
+                        ts = int(self.ig['sd_timestamp'].value * 10)
                          # timestamp is in centiseconds since epoch (40 bit spead limitation)
                         if isinstance(self.storage, SignalDisplayStore2):
                             self.storage.add_data2(ts,  self.ig['sd_data'].value.astype(np.float32).view(np.complex64).swapaxes(0,1)[:,:,0] if hasdata else None, \
